Refactor circuit_utils.go and proving_keys_utils.go
