+++ light-prover/prover/circuit_utils.go
@@ -1,28 +1,14 @@
 package prover
 
 import (
-	"fmt"
-	"light/light-prover/logging"
 	"light/light-prover/prover/poseidon"
 	"math/big"
-	"os"
 
-	"github.com/consensys/gnark-crypto/ecc"
 	"github.com/consensys/gnark/backend/groth16"
 	"github.com/consensys/gnark/constraint"
 	"github.com/consensys/gnark/frontend"
 
 	"github.com/reilabs/gnark-lean-extractor/v2/abstractor"
-)
-
-type RunMode string
-
-const (
-	Forester     RunMode = "forester"
-	ForesterTest RunMode = "forester-test"
-	Rpc          RunMode = "rpc"
-	Full         RunMode = "full"
-	FullTest     RunMode = "full-test"
 )
 
 type Proof struct {
@@ -228,136 +214,4 @@ func (gadget MerkleRootUpdateGadget) DefineGadget(api frontend.API) interface{} 
 	})
 
 	return newRoot
-}
-
-// Trusted setup utility functions
-// Taken from: https://github.com/bnb-chain/zkbnb/blob/master/common/prove/proof_keys.go#L19
-func LoadProvingKey(filepath string) (pk groth16.ProvingKey, err error) {
-	logging.Logger().Info().Msg("start reading proving key")
-	pk = groth16.NewProvingKey(ecc.BN254)
-	f, _ := os.Open(filepath)
-	_, err = pk.ReadFrom(f)
-	if err != nil {
-		return pk, fmt.Errorf("read file error")
-	}
-	err = f.Close()
-	if err != nil {
-		return nil, err
-	}
-	return pk, nil
-}
-
-// Taken from: https://github.com/bnb-chain/zkbnb/blob/master/common/prove/proof_keys.go#L32
-func LoadVerifyingKey(filepath string) (verifyingKey groth16.VerifyingKey, err error) {
-	logging.Logger().Info().Msg("start reading verifying key")
-	verifyingKey = groth16.NewVerifyingKey(ecc.BN254)
-	f, _ := os.Open(filepath)
-	_, err = verifyingKey.ReadFrom(f)
-	if err != nil {
-		return verifyingKey, fmt.Errorf("read file error")
-	}
-	err = f.Close()
-	if err != nil {
-		return nil, err
-	}
-
-	return verifyingKey, nil
-}
-func GetKeys(keysDir string, runMode RunMode, circuits []string) []string {
-	var keys []string
-
-	var inclusionKeys []string = []string{
-		keysDir + "inclusion_26_1.key",
-		keysDir + "inclusion_26_2.key",
-		keysDir + "inclusion_26_3.key",
-		keysDir + "inclusion_26_4.key",
-		keysDir + "inclusion_26_8.key",
-	}
-
-	var nonInclusionKeys []string = []string{
-		keysDir + "non-inclusion_26_1.key",
-		keysDir + "non-inclusion_26_2.key",
-	}
-
-	var combinedKeys []string = []string{
-		keysDir + "combined_26_1_1.key",
-		keysDir + "combined_26_1_2.key",
-		keysDir + "combined_26_2_1.key",
-		keysDir + "combined_26_2_2.key",
-		keysDir + "combined_26_3_1.key",
-		keysDir + "combined_26_3_2.key",
-		keysDir + "combined_26_4_1.key",
-		keysDir + "combined_26_4_2.key",
-	}
-
-	var appendKeys []string = []string{
-		keysDir + "append_26_1.key",
-		keysDir + "append_26_10.key",
-		keysDir + "append_26_100.key",
-		keysDir + "append_26_500.key",
-		keysDir + "append_26_1000.key",
-	}
-	var updateKeys []string = []string{
-		keysDir + "update_26_1.key",
-		keysDir + "update_26_10.key",
-		keysDir + "update_26_100.key",
-		keysDir + "update_26_500.key",
-		keysDir + "update_26_1000.key",
-	}
-
-	var appendTestKeys []string = []string{
-		keysDir + "append_26_10.key",
-	}
-	var updateTestKeys []string = []string{
-		keysDir + "update_26_10.key",
-	}
-
-	switch runMode {
-	case Forester: // inclusion + non-inclusion
-		keys = append(keys, inclusionKeys...)
-		keys = append(keys, nonInclusionKeys...)
-	case ForesterTest: // append-test + update-test
-		keys = append(keys, appendTestKeys...)
-		keys = append(keys, updateTestKeys...)
-	case Rpc: // inclusion + non-inclusion + combined
-		keys = append(keys, inclusionKeys...)
-		keys = append(keys, nonInclusionKeys...)
-		keys = append(keys, combinedKeys...)
-	case Full: // inclusion + non-inclusion + combined + append + update
-		keys = append(keys, inclusionKeys...)
-		keys = append(keys, nonInclusionKeys...)
-		keys = append(keys, combinedKeys...)
-		keys = append(keys, appendKeys...)
-		keys = append(keys, updateKeys...)
-	case FullTest: // inclusion + non-inclusion + combined + append-test + update-test
-		keys = append(keys, inclusionKeys...)
-		keys = append(keys, nonInclusionKeys...)
-		keys = append(keys, combinedKeys...)
-		keys = append(keys, appendTestKeys...)
-		keys = append(keys, updateTestKeys...)
-	}
-
-	fmt.Println("Keys: ", keys)
-	fmt.Println("Circuits: ", circuits)
-
-	for _, circuit := range circuits {
-		switch circuit {
-		case "inclusion":
-			keys = append(keys, inclusionKeys...)
-		case "non-inclusion":
-			keys = append(keys, nonInclusionKeys...)
-		case "combined":
-			keys = append(keys, combinedKeys...)
-		case "append":
-			keys = append(keys, appendKeys...)
-		case "update":
-			keys = append(keys, updateKeys...)
-		case "append-test":
-			keys = append(keys, appendTestKeys...)
-		case "update-test":
-			keys = append(keys, updateTestKeys...)
-		}
-	}
-
-	return keys
 }
